https://github.com/nextic/IC/pull/603

[author: paolafer]

This PR fixes an issue with the way numpy builds histograms, which we
use to create voxels.
@mmkekic found out that changing the precision of the hit positions to
float64 solved this particular problem. I've also added a test to show
the problem in one particular event.

[reviewer: jjgomezcadenas]

The problem seems identified and fixed, with a test demonstrating that
it works.

+++ invisible_cities/reco/paolina_functions.py
@@ -55,7 +55,7 @@ def voxelize_hits(hits             : Sequence[BHit],
     voxel_edges_lo -= eps
     voxel_edges_hi += eps
 
-    hit_positions = np.array([h.pos for h in hits])
+    hit_positions = np.array([h.pos for h in hits]).astype('float64')
     hit_energies  =          [h.E   for h in hits]
     E, edges = np.histogramdd(hit_positions,
                               bins    = number_of_voxels,
